Move translations to TranslationForWorkSheet.js file
