+++ frontend/src/TranslationForWorkSheet.js
@@ -1,0 +1,8 @@
+const typeOfWork = {
+  INSTALLATION: 'telepítés',
+  REPAIR: 'javítás',
+  MAINTENANCE: 'karbantartás',
+  OTHER: 'egyéb',
+}
+
+export { typeOfWork }
